requirements.txt
tornado==4.0.2
python-simple-hipchat==0.3.3